Statement Update
Added property tax and business insurance to the building, so that statements can draw from that instead of having them as expenses.
Corrected an issue with getting the right property_id to query
Fixed bug from trying to get the column "date" when it is called "date_incurred"
Added "Garbage" to the UtilitiesType enum
Added more expenses to the database test.