Pathing and Testing Update
#6
Pathing Settings added, which have a default value, but will be set by the user.
Pass these settings into the pdf_formatting. Changed the file creation to (for now) write a file to that path named "statement_" and the first and last names of the tenant.
Also separated out testing for now, honestly for no reason.
Also renamed the parameters of write_with_printpdf so it has full names instead of letters
Lastly, added the clone macro to stuff, because it makes things easier. Especially expenses make sense to clone around, as only the values are important, no need to keep a real reference afaik.