Expense Backend
Lots going on.
Full implementation of adding expenses from slint.
There is the callback setup in main, which defines what happens immediately, and what is passed off to the expense worker, which does its own work in another thread.
My issues in the past were apparently related to two issues
1, the On_{Name} callback sets up what WILL happen. It's more like a listener subscription. I had thought it was more like polling, where it would halt execution of code waiting for the callback to return, and I would have to loop over it
2, the upgrade_in_event_loop function doesn't work in other threads. It MUST happen in the same thread where the slint component is built.

Tokio was added for threading